# vegro-hr/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Payroll API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/lead-capture": {
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Capture lead",
                "description": "Capture demo waitlist leads",
                "operationId": "leadCapture",
                "requestBody": {
                    "description": "Lead data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jane Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jane@company.com"
                                    },
                                    "company": {
                                        "type": "string",
                                        "example": "Acme Inc"
                                    },
                                    "message": {
                                        "type": "string",
                                        "example": "We want payroll and reporting demo"
                                    },
                                    "source": {
                                        "type": "string",
                                        "example": "landing-page"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Lead captured successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Lead captured"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/tax-profiles": {
            "get": {
                "tags": [
                    "Tax Profiles"
                ],
                "summary": "List all tax profiles",
                "description": "Get all tax profiles",
                "operationId": "listTaxProfiles",
                "responses": {
                    "200": {
                        "description": "Tax profiles retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Tax Profiles"
                ],
                "summary": "Create tax profile",
                "description": "Create a new tax profile",
                "operationId": "createTaxProfile",
                "requestBody": {
                    "description": "Tax profile data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "country_code",
                                    "currency",
                                    "paye_bands"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Kenya PAYE (Monthly)"
                                    },
                                    "country_code": {
                                        "type": "string",
                                        "example": "KE"
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "KES"
                                    },
                                    "paye_bands": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Tax profile created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/tax-profiles/{taxProfile}": {
            "get": {
                "tags": [
                    "Tax Profiles"
                ],
                "summary": "Get tax profile",
                "description": "Get a specific tax profile",
                "operationId": "showTaxProfile",
                "parameters": [
                    {
                        "name": "taxProfile",
                        "in": "path",
                        "description": "Tax profile ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tax profile retrieved successfully"
                    },
                    "404": {
                        "description": "Tax profile not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Tax Profiles"
                ],
                "summary": "Update tax profile",
                "description": "Update a tax profile",
                "operationId": "updateTaxProfile",
                "requestBody": {
                    "description": "Tax profile data",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Tax profile updated successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tax Profiles"
                ],
                "summary": "Delete tax profile",
                "description": "Delete a tax profile",
                "operationId": "deleteTaxProfile",
                "responses": {
                    "200": {
                        "description": "Tax profile deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}